refactor(appkit): split metric runtime modules
